+++ build_daemon/lib/daemon_builder.dart
@@ -4,6 +4,8 @@
 
 import 'dart:async';
 
+import 'package:watcher/src/watch_event.dart';
+
 import 'data/build_status.dart';
 import 'data/server_log.dart';
 
@@ -12,7 +14,8 @@ class DaemonBuilder {
 
   Stream<ServerLog> get logs => Stream.empty();
 
-  Future<void> build(Set<String> targets, Set<String> logToPaths) async {}
+  Future<void> build(Set<String> targets, Set<String> logToPaths,
+      Iterable<WatchEvent> changes) async {}
 
   Future<void> stop() async {}
 }